refactor: refactor folder locations
