+++ src/components/DeleteProjectModal.tsx
@@ -2,7 +2,7 @@
 
 import { useActionState, useEffect } from 'react';
 import { useFormStatus } from 'react-dom';
-import { deleteProject } from '../actions/project';
+import { deleteProject } from '@/features/project/actions/project';
 import styles from './ProjectModal.module.css';
 
 interface DeleteProjectModalProps {
