+++ src/components/UpdateProjectModal.tsx
@@ -1,15 +1,20 @@
 'use client';
 
+import { useEffect } from 'react';
 import { useActionState } from 'react';
 import { useFormStatus } from 'react-dom';
-import { createProject } from '../actions/project';
+import { updateProject } from '@/features/project/actions/project';
 import styles from './ProjectModal.module.css';
-import { useEffect } from 'react';
 
-interface ProjectModalProps {
+interface UpdateProjectModalProps {
   isOpen: boolean;
   onClose: () => void;
-  onProjectCreated: () => void;
+  onProjectUpdated: () => void;
+  project: {
+    id: string;
+    title: string;
+    description: string;
+  };
 }
 
 const initialState = {
@@ -26,27 +31,28 @@ function SubmitButton() {
       className={styles.submitButton}
       disabled={pending}
     >
-      {pending ? 'Creating...' : 'Create Project'}
+      {pending ? 'Updating...' : 'Update Project'}
     </button>
   );
 }
 
-export default function ProjectModal({ isOpen, onClose, onProjectCreated }: ProjectModalProps) {
-  const [state, formAction] = useActionState(createProject, initialState);
+export default function UpdateProjectModal({ isOpen, onClose, onProjectUpdated, project }: UpdateProjectModalProps) {
+  const [state, formAction] = useActionState(updateProject, initialState);
 
   useEffect(() => {
     if (state.success) {
-      onProjectCreated();
+      onProjectUpdated();
     }
-  }, [state.success, onProjectCreated]);
+  }, [state.success, onProjectUpdated]);
 
   if (!isOpen) return null;
 
   return (
     <div className={styles.modalOverlay} onClick={onClose}>
       <div className={styles.modal} onClick={e => e.stopPropagation()}>
-        <h2 className={styles.modalTitle}>Create New Project</h2>
+        <h2 className={styles.modalTitle}>Update Project</h2>
         <form action={formAction} className={styles.form}>
+          <input type="hidden" name="id" value={project.id} />
           {state.error && !state.success && (
             <div className={styles.error}>{state.error}</div>
           )}
@@ -60,6 +66,7 @@ export default function ProjectModal({ isOpen, onClose, onProjectCreated }: Proj
               name="title"
               className={styles.input}
               placeholder="e.g. Marketing Campaign"
+              defaultValue={project.title}
               required
             />
           </div>
@@ -72,6 +79,7 @@ export default function ProjectModal({ isOpen, onClose, onProjectCreated }: Proj
               name="description"
               className={styles.textarea}
               placeholder="e.g. This project will handle our Q4 marketing initiatives"
+              defaultValue={project.description}
               required
             />
           </div>
